requirements.txt
streamlit
pandas
openpyxl
plotly
scikit-learn
pillow
requests